+++ assets/js/hooks/index.js
@@ -18,6 +18,8 @@ import Session from "./session";
 import TextareaAutosize from "./textarea_autosize";
 import Timer from "./timer";
 import UserForm from "./user_form";
+import UtcDateTimeInput from "./utc_datetime_input";
+import UtcTimeInput from "./utc_time_input";
 import VirtualizedLines from "./virtualized_lines";
 
 export default {
@@ -41,5 +43,7 @@ export default {
   TextareaAutosize,
   Timer,
   UserForm,
+  UtcDateTimeInput,
+  UtcTimeInput,
   VirtualizedLines,
 };