+++ lib/livebook_web/live/learn_live.ex
@@ -116,10 +116,6 @@ defmodule LivebookWeb.LearnLive do
   end
 
   @impl true
-  def handle_params(%{"slug" => "new"}, _url, socket) do
-    {:noreply, create_session(socket)}
-  end
-
   def handle_params(%{"slug" => slug}, _url, socket) do
     {notebook, files} = Learn.notebook_by_slug!(slug)
     {:noreply, create_session(socket, notebook: notebook, files_source: {:inline, files})}