+++ lib/livebook_web/live/session_live.ex
@@ -305,10 +305,10 @@ defmodule LivebookWeb.SessionLive do
                   </a>
                 </.menu_item>
                 <.menu_item variant={:danger}>
-                  <.link navigate={~p"/home/sessions/#{@session.id}/close"} role="menuitem">
+                  <button role="menuitem" phx-click="close_session">
                     <.remix_icon icon="close-circle-line" />
                     <span>Close</span>
-                  </.link>
+                  </button>
                 </.menu_item>
               </.menu>
             </div>
@@ -1413,6 +1413,10 @@ defmodule LivebookWeb.SessionLive do
     {:noreply, create_session(socket, notebook: notebook, files_source: {:dir, files_dir})}
   end
 
+  def handle_event("close_session", %{}, socket) do
+    {:noreply, confirm_close_session(socket, socket.assigns.session, redirect_to: ~p"/")}
+  end
+
   def handle_event("star_notebook", %{}, socket) do
     data = socket.private.data
     Livebook.NotebookManager.add_starred_notebook(data.file, data.notebook.name)
@@ -1970,7 +1974,7 @@ defmodule LivebookWeb.SessionLive do
   end
 
   defp insert_cell_below(socket, params) do
-    {type, attrs} = cell_type_and_attrs_from_params(params)
+    {type, attrs} = cell_type_and_attrs_from_params(params, socket)
 
     with {:ok, section, index} <-
            section_with_next_index(
@@ -1984,13 +1988,22 @@ defmodule LivebookWeb.SessionLive do
     socket
   end
 
-  defp cell_type_and_attrs_from_params(%{"type" => "markdown"}), do: {:markdown, %{}}
+  defp cell_type_and_attrs_from_params(%{"type" => "markdown"}, _socket), do: {:markdown, %{}}
 
-  defp cell_type_and_attrs_from_params(%{"type" => "code", "language" => language})
-       when language in ["elixir", "erlang"],
-       do: {:code, %{language: String.to_atom(language)}}
+  defp cell_type_and_attrs_from_params(%{"type" => "code"} = params, socket) do
+    language =
+      case params["language"] do
+        language when language in ["elixir", "erlang"] ->
+          String.to_atom(language)
 
-  defp cell_type_and_attrs_from_params(%{"type" => "diagram"}) do
+        _ ->
+          socket.private.data.notebook.default_language
+      end
+
+    {:code, %{language: language}}
+  end
+
+  defp cell_type_and_attrs_from_params(%{"type" => "diagram"}, _socket) do
     source = """
     <!-- Learn more at https://mermaid-js.github.io/mermaid -->
 
@@ -2006,7 +2019,7 @@ defmodule LivebookWeb.SessionLive do
     {:markdown, %{source: source}}
   end
 
-  defp cell_type_and_attrs_from_params(%{"type" => "image", "url" => url}) do
+  defp cell_type_and_attrs_from_params(%{"type" => "image", "url" => url}, _socket) do
     source = "![](#{url})"
 
     {:markdown, %{source: source}}