+++ lib/livebook_web/router.ex
@@ -60,8 +60,6 @@ defmodule LivebookWeb.Router do
       pipe_through [:browser, :auth]
 
       live "/", HomeLive, :page
-      live "/home/sessions/:session_id/close", HomeLive, :close_session
-      live "/home/sessions/edit_sessions/:action", HomeLive, :edit_sessions
 
       live "/open/:tab", OpenLive, :page
 
@@ -104,6 +102,7 @@ defmodule LivebookWeb.Router do
     scope "/", LivebookWeb do
       pipe_through [:browser, :auth]
 
+      live "/new", HomeLive, :public_new_notebook
       live "/import", OpenLive, :public_import
       live "/open", OpenLive, :public_open
     end